+++ libs/sdk/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @langchain/langgraph-sdk
 
+## 0.0.95
+
+### Patch Changes
+
+- 39cc88f: useStream should receive and merge messages from subgraphs
+- c1ddda1: Fix fetching state with `fetchStateHistory: false` causing crash if thread is empty
+
+## 0.0.94
+
+### Patch Changes
+
+- 11e95e0: Add isStudioUser for custom auth
+
+## 0.0.93
+
+### Patch Changes
+
+- d53c891: Fix useStream race condition when flushing messages
+
 ## 0.0.92
 
 ### Patch Changes